Added code and dependencies for attractor tool
I created this tool as a project for a class after my professor directed
me to the library. It takes a LogicalModel and identifies every
attractor, regardless of length. The algorithm implemented makes use of
a SAT solver, prompting the maven dependency for Sat4j.

There is only one constraint on the algorithm, that Sat4j can spend no
more than 30 seconds on satisfiying a single boolean expression. This
does not limit the runtime of the algorithm as a whole. Every time Sat4j
finds a solution to the provided expression the expression is modified,
whether Sat4j found an attractor or not. For a few randomly generated
test cases of 500 nodes, each with 2 inputs, the algorithm would run for
half an hour or more.

Additionally, a few of the test cases caused a 4GB heap to overflow.
This can be prevented if desired by limiting the value of a variable
noted by a comment.